Move and rename exceptions.py
