+++ server/internal/exceptions.py
