+++ server/manage.py
@@ -4,9 +4,6 @@ from crontab import CronTab
 from flask.ext.script import Manager
 from flask.ext.script import Server
 
-from custom_exceptions import ServerAlreadyExistsError
-from custom_exceptions import UserAlreadyExistsError
-
 from main import app
 
 manager = Manager(app)
@@ -22,6 +19,8 @@ SCAN_COMMAND = '/usr/bin/python /vagrant/app/manage.py scan'
 def setup_dev():
     from internal import server
     from internal import user
+    from internal.exceptions import ServerAlreadyExistsError
+    from internal.exceptions import UserAlreadyExistsError
 
     try:
         server.create_server('ubuntu', 'vinz-ubuntu.student.iastate.edu')
